pyproject.toml
[tool.poetry]
name = "dsba-6156-search-engine"
version = "0.1.0"
description = "A neural search engine created for DSBA 6156: Applied Machine Learning"
authors = ["Kyle McLester <kmcleste@uncc.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.13.0"
farm-haystack = {extras = ["only-faiss"], version = "^1.9.0"}
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pytest = "^7.1.3"
watchdog = "^2.1.9"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
numpy = "^1.23.3"
Cython = "^0.29.32"
scipy = "^1.9.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
black = "22.6.0"
flake8 = "^5.0.4"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
